Change withShownDefault to accept a function
It's common to use this function on a type with a custom "show-like"
function. Under such usage, the ergonomics are not ideal:

```hs
withShownDefault defaultFoo (showFoo defaultFoo)
```

This function is more ergonomic:

```hs
withShownDefault showFoo defaultFoo
```

As an example, the simplified `withDefault = withShownDefault show`
reads very clearly. The previous behavior can also be recovered easily
with `const`.